// web/src/features/dashboard/types.ts
export interface QuotaDataItem {
  id?: number
  user_id?: number
  username?: string
  token_id?: number
  token_name?: string
  model_name?: string
  created_at: number
  token_used?: number
  count?: number
  quota?: number
}